KAN-103 refactor: add types dir in planner and modify typo
