+++ src/app/planner/[id]/components/planner-synopsis-form-container/PlannerSynopsisFormContainer.module.scss
@@ -2,7 +2,7 @@
 @use "@styles/typography" as typo;
 @use "@styles/colors" as color;
 
-.synopsys-form-wrapper {
+.synopsis-form-wrapper {
   @include mixin.flex();
 
   width: 100%;
