+++ src/app/planner/[id]/components/planner-synopsis-form/PlannerSynopsisForm.tsx
@@ -12,7 +12,7 @@ export default function PlannerSynopsisForm() {
     <div className={cx('synopsis-form')}>
       <div className={cx('synopsis-form__title')}>시놉시스</div>
       <Dropdown
-        name="genre"
+        name="synopsis.genre"
         type="outlined"
         placeholder="장르"
         label="장르"
@@ -35,7 +35,7 @@ export default function PlannerSynopsisForm() {
         isRequired={true}
       />
       <Dropdown
-        name="length"
+        name="synopsis.length"
         type="outlined"
         placeholder="분량"
         label="분량"
@@ -51,13 +51,13 @@ export default function PlannerSynopsisForm() {
         ]}
         isRequired={false}
       />
-      <TextField name="purpose" label="기획 의도" />
+      <TextField name="synopsis.purpose" label="기획 의도" />
       <TextField
-        name="logline"
+        name="synopsis.logline"
         label="로그 라인"
         options={{ required: { value: true, message: 'required' } }}
       />
-      <TextField name="example" label="예시 문장" />
+      <TextField name="synopsis.example" label="예시 문장" />
     </div>
   )
 }
