+++ src/app/planner/[id]/types/PlannerTab.ts
@@ -1,0 +1,4 @@
+export type PlannerTabType = {
+  label: string
+  value: string
+}
